unified token count. unified temperatures (all 0.0)
