+++ app/models/alert_prompt.py
@@ -47,6 +47,7 @@ class AlertPrompt(Base):
     created_at = Column(DateTime, nullable=False, default=datetime.now())
     expires_at = Column(DateTime, nullable=False)
     llm_model = Column(String, nullable=False)
+    max_datetime = Column(DateTime, nullable=False)
 
     # Relationships
     user = relationship("AgentController", back_populates="alert_prompts")
