+++ app/tasks/llm_apis/ollama.py
@@ -8,6 +8,8 @@ client = OpenAI(
     api_key='ollama', # required, but unused
 )
 
+ollama_temperature = 0.0
+
 async def get_nomic_embeddings(text: str):
     embeddings = client.embeddings.create(
         model="nomic-embed-text",
@@ -20,7 +22,7 @@ async def get_ollama_validation(alert_prompt: str, alert_parsed_intent: str) -> 
     full_prompt = get_validation_prompt(alert_prompt, alert_parsed_intent)    
     response = client.chat.completions.create(
         model="llama3.1",
-        temperature=0.5,
+        temperature=ollama_temperature,
         messages=[
             {"role": "user", "content": full_prompt},
         ],
@@ -34,7 +36,7 @@ async def get_ollama_verification(alert_prompt: str, alert_parsed_intent: str, d
     full_prompt = get_verification_prompt(alert_prompt, alert_parsed_intent, document)    
     response = client.chat.completions.create(
         model="llama3.1",
-        temperature=0.5,
+        temperature=ollama_temperature,
         messages=[
             {"role": "user", "content": full_prompt},
         ],
@@ -48,7 +50,7 @@ async def get_ollama_alert_generation(alert_parsed_intent: str, document: str, e
     full_prompt = get_generation_prompt(alert_parsed_intent, document, example_response)    
     response = client.chat.completions.create(
         model="llama3.1",
-        temperature=0.5,
+        temperature=ollama_temperature,
         messages=[
             {"role": "user", "content": full_prompt},
         ],
